+++ tezt/tests/main.ml
@@ -66,6 +66,7 @@ let () =
   Mempool.register ~protocols:[Alpha] ;
   (* Tests that are protocol-independent.
      They do not take a protocol as a parameter and thus need to be registered only once. *)
+  Light.register_protocol_independent () ;
   P2p.register_protocol_independent () ;
   Mockup.register_protocol_independent () ;
   Bootstrap.register_protocol_independent () ;